% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow}
\alias{arrow}
\title{A arrow plot using R}
\usage{
arrow(color, lty)
}
\arguments{
\item{color}{the color of arrow (character)}

\item{lty}{the type of arrow lines (integer)}
}
\value{
a plot of a arrow
}
\description{
A arrow plot using R
}
\examples{
\dontrun{
myarrow <- arrow("red",2)
myheart
}
}
