% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brokeheart.R
\name{brokenheart}
\alias{brokenheart}
\title{A broken heart plot using R}
\usage{
brokenheart(color, words)
}
\arguments{
\item{color}{the color of heart (character)}

\item{words}{the greeting (character)}
}
\value{
a plot of a broken heart
}
\description{
A  broken heart plot using R
}
\examples{
\dontrun{
myheart <- brokenheart("red","I am an orphan")
myheart
}
}
