% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\title{The overall composition using R}
\usage{
design(style, words, lty)
}
\arguments{
\item{style}{the style of plot (integer)}

\item{words}{the greeting (character)}

\item{lty}{the type of line (integer)}
}
\value{
a plot of some hearts
}
\description{
The overall composition using R
}
\examples{
\dontrun{
myplot <- design(1,"Merry Christmas !",2)
myplot
}
}
