% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{heart}
\alias{heart}
\title{A heart plot using R}
\usage{
heart(color, lty)
}
\arguments{
\item{color}{the color of heart (character)}

\item{lty}{the type of heart lines (integer)}
}
\value{
a plot of a heart
}
\description{
A heart plot using R
}
\examples{
\dontrun{
myheart <- heart("red",2)
myheart
}
}
